+++ src/Workflow/Entity/Database/DatabaseEntityRepository.php
@@ -15,6 +15,7 @@ declare(strict_types=1);
 
 namespace Netzmacht\ContaoWorkflowBundle\Workflow\Entity\Database;
 
+use ArrayObject;
 use Doctrine\DBAL\Connection;
 use Netzmacht\ContaoWorkflowBundle\Workflow\Exception\UnsupportedEntity;
 use Netzmacht\Workflow\Data\EntityRepository;
@@ -57,7 +58,7 @@ final class DatabaseEntityRepository implements EntityRepository
      *
      * @throws \InvalidArgumentException When an entity could not be found.
      */
-    public function find($entityId): array
+    public function find($entityId): ArrayObject
     {
         $statement = $this->connection->createQueryBuilder()
             ->select('*')
@@ -69,7 +70,7 @@ final class DatabaseEntityRepository implements EntityRepository
         $row = $statement->fetch(\PDO::FETCH_ASSOC);
 
         if (is_array($row)) {
-            return $row;
+            return new ArrayObject($row);
         }
 
         throw new \InvalidArgumentException(sprintf('Could not find entity "%s"', $entityId));
@@ -92,14 +93,14 @@ final class DatabaseEntityRepository implements EntityRepository
      */
     public function add($entity): void
     {
-        if (!is_array($entity)) {
+        if (!$entity instanceof ArrayObject) {
             throw UnsupportedEntity::forEntity($entity);
         }
 
         if (isset($entity['id'])) {
-            $this->connection->update($this->providerName, $entity, ['id' => $entity['id']]);
+            $this->connection->update($this->providerName, $entity->getArrayCopy(), ['id' => $entity['id']]);
         } else {
-            $this->connection->insert($this->providerName, $entity);
+            $this->connection->insert($this->providerName, $entity->getArrayCopy());
         }
     }
 
@@ -110,7 +111,7 @@ final class DatabaseEntityRepository implements EntityRepository
      */
     public function remove($entity): void
     {
-        if (!is_array($entity)) {
+        if (!$entity instanceof ArrayObject) {
             throw UnsupportedEntity::forEntity($entity);
         }